requirements.txt
django
pytesseract   
djangorestframework
markdown
django-filter
celery
django-celery